Clean-up after IAM policy revamp
Permissions to S3 are not in managed IAM policies anymore.
They're in the user role `s3-access` inline policy.

This means we can delete a bunch of functions which attach/detach policies,
etc...
 - `aws.attach_policy_to_role()`
 - `aws.create_policy()`
 - `aws.delete_policy()`
 - `aws.detach_policy_from_entities()`
 - `aws.detach_policy_from_group()`
 - `aws.detach_policy_from_user()`
 - `models.S3Bucket.aws_delete()`
 - `services.create_bucket_policies()`
 - `services.delete_bucket_policies()`
 - `services.get_policy_document()`
 - `views.S3BucketViewSet.perform_destroy()`

I also moved `_policy_name()` and `_policy_aen()` from the `services` module
into the `migrate_s3_access_roles_to_inline_policies` migration which
is the only place using them. This is not strictly necessary in theory
as I already run this migration but I preferred to keep this migration
functioning.
