+++ control_panel_api/services.py
@@ -12,8 +12,6 @@ from control_panel_api.aws import (
 
 S3_POLICY_NAME = 's3-access'
 
-READWRITE = 'readwrite'
-READONLY = 'readonly'
 
 logger = logging.getLogger(__name__)
 
@@ -38,24 +36,6 @@ def ignore_aws_exceptions(func):
             logger.error(f"Caught aws exception and ignored: {e}")
 
     return inner
-
-
-def _policy_name(bucket_name, readwrite=False):
-    """
-    Prefix the policy name with bucket name, postfix with access level
-    eg: dev-james-readwrite
-    """
-    return "{}-{}".format(bucket_name, READWRITE if readwrite else READONLY)
-
-
-def _policy_arn(bucket_name, readwrite=False):
-    """
-    Return full bucket policy arn
-    eg: arn:aws:iam::1337:policy/bucketname-readonly
-    """
-    return "{}:policy/{}".format(
-        settings.IAM_ARN_BASE,
-        _policy_name(bucket_name, readwrite))
 
 
 @ignore_aws_exceptions
@@ -110,59 +90,6 @@ def delete_role(role_name):
     aws.delete_role(role_name)
 
 
-def get_policy_document(bucket_name_arn, readwrite):
-    statements = [
-        {
-            "Sid": "ListBucketsInConsole",
-            "Effect": "Allow",
-            "Action": [
-                "s3:GetBucketLocation",
-                "s3:ListAllMyBuckets"
-            ],
-            "Resource": "arn:aws:s3:::*"
-        },
-        {
-            "Sid": "ListObjects",
-            "Action": [
-                "s3:ListBucket"
-            ],
-            "Effect": "Allow",
-            "Resource": [bucket_name_arn],
-        },
-        {
-            "Sid": "ReadObjects",
-            "Action": [
-                "s3:GetObject",
-                "s3:GetObjectAcl",
-                "s3:GetObjectVersion",
-            ],
-            "Effect": "Allow",
-            "Resource": "{}/*".format(bucket_name_arn)
-        },
-    ]
-
-    if readwrite:
-        statements.append(
-            {
-                "Sid": "UpdateRenameAndDeleteObjects",
-                "Action": [
-                    "s3:DeleteObject",
-                    "s3:DeleteObjectVersion",
-                    "s3:PutObject",
-                    "s3:PutObjectAcl",
-                    "s3:RestoreObject",
-                ],
-                "Effect": "Allow",
-                "Resource": "{}/*".format(bucket_name_arn)
-            }
-        )
-
-    return {
-        "Version": "2012-10-17",
-        "Statement": statements,
-    }
-
-
 @ignore_aws_exceptions
 def create_bucket(bucket_name):
     aws.create_bucket(
@@ -174,36 +101,6 @@ def create_bucket(bucket_name):
         target_bucket=settings.LOGS_BUCKET_NAME,
         target_prefix=f"{bucket_name}/")
     aws.put_bucket_encryption(bucket_name)
-
-
-@ignore_aws_exceptions
-def create_bucket_policies(bucket_name, bucket_arn):
-    """Create readwrite and readonly policies for s3 bucket"""
-    readwrite = True
-    policy_name = _policy_name(bucket_name, readwrite)
-    policy_document = get_policy_document(bucket_arn, readwrite)
-
-    aws.create_policy(policy_name, policy_document)
-
-    readwrite = False
-    policy_name = _policy_name(bucket_name, readwrite)
-    policy_document = get_policy_document(bucket_arn, readwrite)
-
-    aws.create_policy(policy_name, policy_document)
-
-
-def delete_bucket_policies(bucket_name):
-    """
-    Delete policy from attached entities first then delete policy, for both
-    policy types
-    """
-    policy_arn_readwrite = _policy_arn(bucket_name, readwrite=True)
-    aws.detach_policy_from_entities(policy_arn_readwrite)
-    aws.delete_policy(policy_arn_readwrite)
-
-    policy_arn_readonly = _policy_arn(bucket_name, readwrite=False)
-    aws.detach_policy_from_entities(policy_arn_readonly)
-    aws.delete_policy(policy_arn_readonly)
 
 
 @contextmanager
@@ -226,7 +123,6 @@ def s3_access_policy(role_name):
 def revoke_bucket_access(bucket_arn, role_name):
     with s3_access_policy(role_name) as policy:
         policy.revoke_access(bucket_arn)
-
 
 
 def grant_bucket_access(bucket_arn, readwrite, role_name):
