+++ control_panel_api/tests/test_aws.py
@@ -18,17 +18,6 @@ class AwsTestCase(TestCase):
         aws.put_bucket_logging('bucketname', 'target_bucket', 'target_prefix')
         aws.client.return_value.put_bucket_logging.assert_called()
 
-    def test_create_policy_json_encoded(self):
-        aws.create_policy('bucketname', {'foo': 'bar'})
-        aws.client.return_value.create_policy.assert_called_with(
-            PolicyName='bucketname',
-            PolicyDocument='{"foo": "bar"}'
-        )
-
-    def test_delete_policy(self):
-        aws.delete_policy('policyarn')
-        aws.client.return_value.delete_policy.assert_called()
-
     def test_get_inline_policy_document_when_found(self):
         role_name = 'test-user-rolename'
         policy_name = 's3-access'
@@ -68,47 +57,9 @@ class AwsTestCase(TestCase):
             PolicyDocument=json.dumps(policy_document),
         )
 
-    def test_detach_policy_from_entities(self):
-        aws.client.return_value.list_entities_for_policy.return_value = {
-            'PolicyRoles': [{'RoleName': 'foo'}],
-            'PolicyGroups': [{'GroupName': 'bar'}],
-            'PolicyUsers': [{'UserName': 'baz'}],
-        }
-
-        aws.detach_policy_from_entities('policyarn')
-
-        aws.client.return_value.list_entities_for_policy.assert_called()
-        aws.client.return_value.detach_role_policy.assert_called_with(
-            RoleName='foo',
-            PolicyArn='policyarn',
-        )
-        aws.client.return_value.detach_group_policy.assert_called_with(
-            GroupName='bar',
-            PolicyArn='policyarn',
-        )
-        aws.client.return_value.detach_user_policy.assert_called_with(
-            UserName='baz',
-            PolicyArn='policyarn',
-        )
-
-    def test_attach_policy_to_role(self):
-        aws.attach_policy_to_role('policyarn', 'rolename')
-        aws.client.return_value.attach_role_policy.assert_called_with(
-            RoleName='rolename',
-            PolicyArn='policyarn',
-        )
-
     def test_detach_policy_from_role(self):
         aws.detach_policy_from_role('policyarn', 'foo')
         aws.client.return_value.detach_role_policy.assert_called()
-
-    def test_detach_policy_from_group(self):
-        aws.detach_policy_from_group('policyarn', 'foo')
-        aws.client.return_value.detach_group_policy.assert_called()
-
-    def test_detach_policy_from_user(self):
-        aws.detach_policy_from_user('policyarn', 'foo')
-        aws.client.return_value.detach_user_policy.assert_called()
 
     def test_create_role(self):
         role_name = "a_role"
